+++ src/server/index.ts
@@ -12,11 +12,11 @@ import helmet from "helmet";
 import morgan from "morgan";
 import multer from "multer";
 import nacl from "tweetnacl";
-import * as uuid from "uuid";
 import winston from "winston";
 
 import { Database } from "../Database";
 import { getUserRouter } from "./user";
+import { getFileRouter } from "./file";
 
 // expiry of regkeys
 export const EXPIRY_TIME = 1000 * 60 * 5;
@@ -38,6 +38,7 @@ export const initApp = (
 ) => {
     // INIT ROUTERS
     const userRouter = getUserRouter(db, log, tokenValidator);
+    const fileRouter = getFileRouter(db, log);
 
     api.use(express.json({ limit: "20mb" }));
     api.use(helmet());
@@ -68,10 +69,6 @@ export const initApp = (
         }
     });
 
-    // user
-    api.use("/user", userRouter);
-
-    // device
     api.get("/device/:id", async (req, res) => {
         const device = await db.retrieveDevice(req.params.id);
 
@@ -82,84 +79,14 @@ export const initApp = (
         }
     });
 
+    api.get("/canary", async (req, res) => {
+        res.send({ canary: process.env.CANARY });
+    });
+
+    api.use("/user", userRouter);
+
     // file
-    api.get("/file/:id", async (req, res) => {
-        const entry = await db.retrieveFile(req.params.id);
-        if (!entry) {
-            res.sendStatus(404);
-        } else {
-            fs.readFile(
-                path.resolve("./files/" + entry.fileID),
-                undefined,
-                async (err, file) => {
-                    if (err) {
-                        log.error("error reading file");
-                        log.error(err);
-                        res.sendStatus(500);
-                    } else {
-                        const typeDetails = await FileType.fromBuffer(file);
-                        // TODO: fix this as well, its bloating the size
-                        const resp: XTypes.HTTP.IFileResponse = {
-                            details: entry,
-                            data: file,
-                        };
-                        if (typeDetails) {
-                            res.set("Content-type", typeDetails.mime);
-                        }
-                        res.send(resp);
-                    }
-                }
-            );
-        }
-    });
-
-    api.post("/file", multer().single("file"), async (req, res) => {
-        const payload: XTypes.HTTP.IFilePayload = req.body;
-
-        if (payload.nonce === "") {
-            res.sendStatus(400);
-            return;
-        }
-
-        const deviceEntry = await db.retrieveDevice(payload.owner);
-        if (!deviceEntry) {
-            log.warn("No device found.");
-            res.send(400);
-            return;
-        }
-
-        const devices = await db.retrieveUserDeviceList(deviceEntry.owner);
-
-        let token: Uint8Array | null = null;
-        for (const device of devices) {
-            const verified = nacl.sign.open(
-                XUtils.decodeHex(payload.signed),
-                XUtils.decodeHex(device.signKey)
-            );
-            if (verified) {
-                token = verified;
-            }
-        }
-        if (!token) {
-            log.warn("Bad signature on token.");
-            res.sendStatus(401);
-            return;
-        }
-
-        const newFile: XTypes.SQL.IFile = {
-            fileID: uuid.v4(),
-            owner: payload.owner,
-            nonce: payload.nonce,
-        };
-
-        // write the file to disk
-        fs.writeFile("files/" + newFile.fileID, req.file.buffer, () => {
-            log.info("Wrote new file " + newFile.fileID);
-        });
-
-        await db.createFile(newFile);
-        res.send(newFile);
-    });
+    api.use("/file", fileRouter);
 
     // avatar
     api.get("/avatar/:userID", async (req, res) => {
@@ -219,10 +146,6 @@ export const initApp = (
             log.warn(err);
             res.sendStatus(500);
         }
-    });
-
-    api.get("/canary", async (req, res) => {
-        res.send({ canary: process.env.CANARY });
     });
 };
